move color to its own faile
