+++ packages/gui/src/components/schemas/box-shadow.tsx
@@ -1,8 +1,9 @@
 import { CheckboxInput } from '../inputs/CheckboxInput'
+import { color } from './color'
 import { joinSchemas } from './joinSchemas'
 import { listSchema } from './list'
 import { objectSchema } from './object'
-import { color, keyword, length } from './primitives'
+import { keyword, length } from './primitives'
 import { DataTypeSchema } from './types'
 
 const inset: DataTypeSchema<boolean> = {
